Fixing Errors
Fixing the return type for the min max function

+++ src/include/MinMax.h
@@ -3,7 +3,7 @@
 #include "move.h"
 #include"Constants.h"
 float evaluate_board(int board[ROWS][COLUMNS], int player);
-int minimax(int board[ROWS][COLUMNS], int depth, bool maximizingPlayer, int player) ;
+float minimax(int board[ROWS][COLUMNS], int depth, bool maximizingPlayer, int player) ;
 struct Move* best_move(int board[ROWS][COLUMNS], int player,int Depth) ;
 
 bool is_game_over(int board[ROWS][COLUMNS]);
